fix: 페이징 수정

+++ projectgooroom/src/main/java/GooRoom/projectgooroom/homepost/dto/HomePostListDto.java
@@ -3,7 +3,7 @@ package GooRoom.projectgooroom.homepost.dto;
 import java.util.List;
 
 public record HomePostListDto(
-        int totalMates,
+        long totalMates,
         List<ListedPostDto> mateList
 ) {
 }
